complete model 3d api, create cart, order api, upgrade upload files

+++ src/main/java/fpt/custome/yourlure/entity/DiscountVoucher.java
@@ -61,7 +61,7 @@ public class DiscountVoucher {
 
     @Nullable
     @Column(name = "minSpentAmount")
-    private Integer minSpentAmount;
+    private Float minSpentAmount;
 
     @Nullable
     @Column(name = "minCheckoutItemsQuantity")
